Agent Naming Convention Fixes
