+++ app/modules/intelligence/agents/agents/integration_test_agent.py
@@ -17,7 +17,7 @@ from app.modules.intelligence.tools.kg_based_tools.get_code_from_probable_node_n
 )
 
 
-class IntegrationTestCrew:
+class IntegrationTestAgent:
     def __init__(self, sql_db, llm, user_id):
         self.openai_api_key = os.getenv("OPENAI_API_KEY")
         self.user_id = user_id
@@ -182,7 +182,7 @@ class IntegrationTestCrew:
         return result
 
 
-async def kickoff_integration_test_crew(
+async def kickoff_integration_test_agent(
     query: str,
     project_id: str,
     node_ids: List[NodeContext],
@@ -215,7 +215,7 @@ async def kickoff_integration_test_crew(
         return node_contexts
 
     node_contexts = extract_unique_node_contexts(graph["graph"]["root_node"])
-    integration_test_agent = IntegrationTestCrew(sql_db, llm, user_id)
+    integration_test_agent = IntegrationTestAgent(sql_db, llm, user_id)
     result = await integration_test_agent.run(
         project_id, node_contexts, query, graph, history
     )
