+++ app/modules/intelligence/agents/chat_agents/lld_chat_agent.py
@@ -18,7 +18,7 @@ from sqlalchemy.orm import Session
 from app.modules.conversations.message.message_model import MessageType
 from app.modules.conversations.message.message_schema import NodeContext
 from app.modules.intelligence.agents.agents_service import AgentsService
-from app.modules.intelligence.agents.crew.rag_crew import kickoff_rag_crew
+from app.modules.intelligence.agents.agents.rag_agent import kickoff_rag_agent
 from app.modules.intelligence.memory.chat_history_service import ChatHistoryService
 from app.modules.intelligence.prompts.classification_prompts import (
     AgentType,
@@ -32,7 +32,7 @@ from app.modules.intelligence.prompts.prompt_service import PromptService
 logger = logging.getLogger(__name__)
 
 
-class LLDAgent:
+class LLDChatAgent:
     def __init__(self, mini_llm, llm, db: Session):
         self.mini_llm = mini_llm
         self.llm = llm
@@ -118,7 +118,7 @@ class LLDAgent:
             citations = []
             if classification == ClassificationResult.AGENT_REQUIRED:
                 rag_start_time = time.time()  # Start timer for RAG agent
-                rag_result = await kickoff_rag_crew(
+                rag_result = await kickoff_rag_agent(
                     query,
                     project_id,
                     [
@@ -233,5 +233,5 @@ class LLDAgent:
                 )
 
         except Exception as e:
-            logger.error(f"Error during QNAAgent run: {str(e)}", exc_info=True)
+            logger.error(f"Error during LLDChatAgent run: {str(e)}", exc_info=True)
             yield f"An error occurred: {str(e)}"
